ml/requirements.txt
# ML Training Dependencies
# Python 3.8+

# Core ML Libraries
torch>=2.0.0
transformers>=4.35.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# OpenAI API (for fine-tuning)
openai>=1.0.0

# Data Processing
datasets>=2.14.0
tokenizers>=0.15.0

# Evaluation
evaluate>=0.4.0
rouge-score>=0.1.2

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Optional: For quantization
# bitsandbytes>=0.41.0

# Optional: For LoRA fine-tuning
# peft>=0.6.0

# Optional: For RLHF
# trl>=0.7.0

